START: Extract Controller class
