+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/AbstractParallelEoSStreamProcessor.java
@@ -6,9 +6,7 @@ package io.confluent.parallelconsumer.internal;
 
 import io.confluent.csid.utils.TimeUtils;
 import io.confluent.parallelconsumer.ParallelConsumer;
-import io.confluent.parallelconsumer.ParallelConsumerException;
 import io.confluent.parallelconsumer.ParallelConsumerOptions;
-import io.confluent.parallelconsumer.PollContextInternal;
 import io.confluent.parallelconsumer.state.WorkManager;
 import io.micrometer.core.instrument.Timer;
 import io.micrometer.core.instrument.simple.SimpleMeterRegistry;
@@ -17,30 +15,16 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerRebalanceListener;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.consumer.MockConsumer;
-import org.apache.kafka.clients.consumer.internals.ConsumerCoordinator;
-import org.apache.kafka.common.TopicPartition;
 import org.slf4j.MDC;
 
-import javax.naming.InitialContext;
-import javax.naming.NamingException;
 import java.io.Closeable;
-import java.lang.reflect.Field;
 import java.time.Clock;
 import java.time.Duration;
-import java.time.Instant;
-import java.util.*;
-import java.util.concurrent.*;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.function.Consumer;
-import java.util.function.Function;
+import java.util.Collection;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.regex.Pattern;
 
-import static io.confluent.csid.utils.BackportUtils.isEmpty;
-import static io.confluent.parallelconsumer.internal.State.*;
-import static java.lang.Boolean.TRUE;
-import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static lombok.AccessLevel.PROTECTED;
 import static lombok.AccessLevel.PUBLIC;
 
@@ -54,8 +38,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
 
     @Getter(PROTECTED)
     protected final ParallelConsumerOptions<K, V> options;
-
-    private final PCModule<K, V> module;
 
     /**
      * Injectable clock for testing
@@ -87,109 +69,15 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         return options.getCommitInterval();
     }
 
-    private Instant lastCommitCheckTime = Instant.now();
-
     @Getter(PROTECTED)
     private final Optional<ProducerManager<K, V>> producerManager;
 
     private final org.apache.kafka.clients.consumer.Consumer<K, V> consumer;
 
-    /**
-     * The pool which is used for running the users' supplied function
-     */
-    protected final ThreadPoolExecutor workerThreadPool;
-
-    private PCWorkerPool<K, V, Object> workerPool;
-
-    private Optional<Future<Boolean>> controlThreadFuture = Optional.empty();
 
     // todo make package level
     @Getter(PUBLIC)
     protected final WorkManager<K, V> wm;
-
-    // todo make private
-    @Getter(PUBLIC)
-    private WorkMailbox<K, V> workMailbox;
-
-    private final BrokerPollSystem<K, V> brokerPollSubsystem;
-
-    /**
-     * Useful for testing async code
-     */
-    private final List<Runnable> controlLoopHooks = new ArrayList<>();
-
-    /**
-     * Reference to the control thread, used for waking up a blocking poll ({@link BlockingQueue#poll}) against a
-     * collection sooner.
-     *
-     * @see #processWorkCompleteMailBox
-     */
-    private Thread blockableControlThread;
-
-    private final OffsetCommitter committer;
-
-    /**
-     * Used to request a commit asap
-     */
-    private final AtomicBoolean commitCommand = new AtomicBoolean(false);
-
-    /**
-     * Multiple of {@link ParallelConsumerOptions#getMaxConcurrency()} to have in our processing queue, in order to make
-     * sure threads always have work to do.
-     */
-    protected final DynamicLoadFactor dynamicExtraLoadFactor;
-
-    /**
-     * If the system failed with an exception, it is referenced here.
-     */
-    private Exception failureReason;
-
-    /**
-     * Time of last successful commit
-     */
-    private Instant lastCommitTime;
-
-    public boolean isClosedOrFailed() {
-        boolean closed = state == State.closed;
-        boolean doneOrCancelled = false;
-        if (this.controlThreadFuture.isPresent()) {
-            Future<Boolean> threadFuture = controlThreadFuture.get();
-            doneOrCancelled = threadFuture.isDone() || threadFuture.isCancelled();
-        }
-        return closed || doneOrCancelled;
-    }
-
-    /**
-     * @return if the system failed, returns the recorded reason.
-     */
-    public Exception getFailureCause() {
-        return this.failureReason;
-    }
-
-    /**
-     * The run state of the controller.
-     *
-     * @see State
-     */
-    @Setter
-    private State state = State.unused;
-
-    /**
-     * Wrapped {@link ConsumerRebalanceListener} passed in by a user that we can also call on events
-     */
-    private Optional<ConsumerRebalanceListener> usersConsumerRebalanceListener = Optional.empty();
-
-    @Getter
-    private int numberOfAssignedPartitions;
-
-    private final RateLimiter queueStatsLimiter = new RateLimiter();
-
-    /**
-     * Control for stepping loading factor - shouldn't step if work requests can't be fulfilled due to restrictions.
-     * (e.g. we may want 10, but maybe there's a single partition and we're in partition mode - stepping up won't
-     * help).
-     */
-    private boolean lastWorkRequestWasFulfilled = false;
 
     private final SimpleMeterRegistry metricsRegistry = new SimpleMeterRegistry();
     private final Timer workRetrievalTimer = metricsRegistry.timer("user.function");
@@ -207,7 +95,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
      */
     protected AbstractParallelEoSStreamProcessor(ParallelConsumerOptions<K, V> newOptions, PCModule<K, V> module) {
         Objects.requireNonNull(newOptions, "Options must be supplied");
-        this.module = module;
 
         options = newOptions;
         this.consumer = options.getConsumer();
@@ -218,77 +105,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         log.info("Confluent Parallel Consumer initialise... groupId: {}, Options: {}",
                 newOptions.getConsumer().groupMetadata().groupId(),
                 newOptions);
-
-
-        this.dynamicExtraLoadFactor = module.dynamicExtraLoadFactor();
-
-        workerThreadPool = setupWorkerPool(newOptions.getMaxConcurrency());
-
-        this.wm = module.workManager();
-
-        this.brokerPollSubsystem = module.brokerPoller();
-
-        if (options.isProducerSupplied()) {
-            this.producerManager = Optional.of(module.producerManager());
-            if (options.isUsingTransactionalProducer())
-                this.committer = this.producerManager.get();
-            else
-                this.committer = this.brokerPollSubsystem;
-        } else {
-            this.producerManager = Optional.empty();
-            this.committer = this.brokerPollSubsystem;
-        }
-
-        this.workMailbox = module.workMailbox();
-    }
-
-    private void validateConfiguration() {
-        options.validate();
-
-        checkGroupIdConfigured(consumer);
-        checkNotSubscribed(consumer);
-        checkAutoCommitIsDisabled(consumer);
-    }
-
-    private void checkGroupIdConfigured(final org.apache.kafka.clients.consumer.Consumer<K, V> consumer) {
-        try {
-            consumer.groupMetadata();
-        } catch (RuntimeException e) {
-            throw new IllegalArgumentException("Error validating Consumer configuration - no group metadata - missing a " +
-                    "configured GroupId on your Consumer?", e);
-        }
-    }
-
-    protected ThreadPoolExecutor setupWorkerPool(int poolSize) {
-        ThreadFactory defaultFactory;
-        try {
-            defaultFactory = InitialContext.doLookup(options.getManagedThreadFactory());
-        } catch (NamingException e) {
-            log.debug("Using Java SE Thread", e);
-            defaultFactory = Executors.defaultThreadFactory();
-        }
-        ThreadFactory finalDefaultFactory = defaultFactory;
-        ThreadFactory namingThreadFactory = r -> {
-            Thread thread = finalDefaultFactory.newThread(r);
-            String name = thread.getName();
-            thread.setName("pc-" + name);
-            return thread;
-        };
-        ThreadPoolExecutor.AbortPolicy rejectionHandler = new ThreadPoolExecutor.AbortPolicy();
-        LinkedBlockingQueue<Runnable> workQueue = new LinkedBlockingQueue<>();
-        return new ThreadPoolExecutor(poolSize, poolSize, 0L, MILLISECONDS, workQueue,
-                namingThreadFactory, rejectionHandler);
-    }
-
-    private void checkNotSubscribed(org.apache.kafka.clients.consumer.Consumer<K, V> consumerToCheck) {
-        if (consumerToCheck instanceof MockConsumer)
-            // disabled for unit tests which don't test rebalancing
-            return;
-        Set<String> subscription = consumerToCheck.subscription();
-        Set<TopicPartition> assignment = consumerToCheck.assignment();
-        if (!subscription.isEmpty() || !assignment.isEmpty()) {
-            throw new IllegalStateException("Consumer subscription must be managed by the Parallel Consumer. Use " + this.getClass().getName() + "#subcribe methods instead.");
-        }
     }
 
     @Override
@@ -317,283 +133,12 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
         consumer.subscribe(pattern, this);
     }
 
-    /**
-     * Nasty reflection to check if auto commit is disabled.
-     * <p>
-     * Other way would be to politely request the user also include their consumer properties when construction, but
-     * this is more reliable in a correctness sense, but brittle in terms of coupling to internal implementation.
-     * Consider requesting ability to inspect configuration at runtime.
-     */
-    private void checkAutoCommitIsDisabled(org.apache.kafka.clients.consumer.Consumer<K, V> consumer) {
-        try {
-            if (consumer instanceof KafkaConsumer) {
-                // Could use Commons Lang FieldUtils#readField - but, avoid needing commons lang
-                Field coordinatorField = KafkaConsumer.class.getDeclaredField("coordinator");
-                coordinatorField.setAccessible(true);
-                ConsumerCoordinator coordinator = (ConsumerCoordinator) coordinatorField.get(consumer); //IllegalAccessException
-
-                if (coordinator == null)
-                    throw new IllegalStateException("Coordinator for Consumer is null - missing GroupId? Reflection broken?");
-
-                Field autoCommitEnabledField = coordinator.getClass().getDeclaredField("autoCommitEnabled");
-                autoCommitEnabledField.setAccessible(true);
-                Boolean isAutoCommitEnabled = (Boolean) autoCommitEnabledField.get(coordinator);
-
-                if (TRUE.equals(isAutoCommitEnabled))
-                    throw new ParallelConsumerException("Consumer auto commit must be disabled, as commits are handled by the library.");
-            } else if (consumer instanceof MockConsumer) {
-                log.debug("Detected MockConsumer class which doesn't do auto commits");
-            } else {
-                // Probably Mockito
-                log.error("Consumer is neither a KafkaConsumer nor a MockConsumer - cannot check auto commit is disabled for consumer type: " + consumer.getClass().getName());
-            }
-        } catch (NoSuchFieldException | IllegalAccessException e) {
-            throw new IllegalStateException("Cannot check auto commit is disabled for consumer type: " + consumer.getClass().getName(), e);
-        }
-    }
-
-
-    private boolean isRecordsAwaitingProcessing() {
-        boolean isRecordsAwaitingProcessing = wm.isRecordsAwaitingProcessing();
-        boolean threadsDone = areMyThreadsDone();
-        log.trace("isRecordsAwaitingProcessing {} || threadsDone {}", isRecordsAwaitingProcessing, threadsDone);
-        return isRecordsAwaitingProcessing || threadsDone;
-    }
-
-
-    /**
-     * Control thread can be blocked waiting for work, but is interruptible. Interrupting it can be useful to inform
-     * that work is available when there was none, to make tests run faster, or to move on to shutting down the
-     * {@link BrokerPollSystem} so that less messages are downloaded and queued.
-     */
-    private void interruptControlThread() {
-        if (blockableControlThread != null) {
-            log.debug("Interrupting {} thread in case it's waiting for work", blockableControlThread.getName());
-            blockableControlThread.interrupt();
-        }
-    }
-
-    private boolean areMyThreadsDone() {
-        if (isEmpty(controlThreadFuture)) {
-            // not constructed yet, will become alive, unless #poll is never called
-            return false;
-        } else {
-            return controlThreadFuture.get().isDone();
-        }
-    }
-
-
-    private static void initWorkerPool(Function userFunctionWrapped, Consumer<Object> callback) {
-        // todo casts
-        Function<PollContextInternal<K, V>, List<Object>> cast = userFunctionWrapped;
-        var runner = FunctionRunner.<K, V, Object>builder()
-                .userFunctionWrapped(cast)
-                .callback(callback)
-                .options(options)
-                .module(module)
-                .workMailbox(workMailbox)
-                .workManager(wm)
-                .build();
-        workerPool = new PCWorkerPool<>(options.getMaxConcurrency(), runner, options);
-    }
 
     /**
      * Useful when testing with more than one instance
      */
     public static void addInstanceMDC(ParallelConsumerOptions<?, ?> options) {
         options.getMyId().ifPresent(id -> MDC.put(MDC_INSTANCE_ID, id));
-    }
-
-
-    /**
-     * If we don't have enough work queued, and the poller is paused for throttling,
-     * <p>
-     * todo move into {@link WorkManager} as it's specific to WM having enough work?
-     */
-    private void maybeWakeupPoller() {
-        if (state == running) {
-            if (!wm.isSufficientlyLoaded() && brokerPollSubsystem.isPausedForThrottling()) {
-                log.debug("Found Poller paused with not enough front loaded messages, ensuring poller is awake (mail: {} vs target: {})",
-                        wm.getNumberOfWorkQueuedInShardsAwaitingSelection(),
-                        options.getTargetAmountOfRecordsInFlight());
-                brokerPollSubsystem.wakeupIfPaused();
-            }
-        }
-    }
-
-    /**
-     * If it's time to commit, and using transactional system, tries to acquire the commit lock.
-     * <p>
-     * Call {@link ProducerManager#preAcquireOffsetsToCommit()} early, to initiate the record sending barrier for this
-     * transaction (so no more records can be sent, before collecting offsets to commit).
-     *
-     * @return true if committing should either way be attempted now
-     */
-    private boolean maybeAcquireCommitLock() throws TimeoutException, InterruptedException {
-        final boolean shouldTryCommitNow = isTimeToCommitNow() && wm.isDirty();
-        // could do this optimistically as well, and only get the lock if it's time to commit, so is not frequent
-        if (shouldTryCommitNow && options.isUsingTransactionCommitMode()) {
-            // get into write lock queue, so that no new work can be started from here on
-            log.debug("Acquiring commit lock pessimistically, before we try to collect offsets for committing");
-            //noinspection OptionalGetWithoutIsPresent - options will already be verified
-            producerManager.get().preAcquireOffsetsToCommit();
-        }
-        return shouldTryCommitNow;
-    }
-
-    private <R> int retrieveAndDistributeNewWorkNew() {
-        var capacity = workerPool.getCapacity(workRetrievalTimer);
-        var work = wm.getWorkIfAvailable(capacity);
-        workerPool.distribute(work);
-        return work.size();
-    }
-
-
-    /**
-     * @return aim to never have the pool queue drop below this
-     */
-    private int getPoolLoadTarget() {
-        return options.getTargetAmountOfRecordsInFlight();
-    }
-
-
-    /**
-     * The amount of time to block poll in this cycle
-     *
-     * @return either the duration until next commit, or next work retry
-     * @see ParallelConsumerOptions#getTargetAmountOfRecordsInFlight()
-     */
-    private Duration getTimeToBlockFor() {
-        // if less than target work already in flight, don't sleep longer than the next retry time for failed work, if it exists - so that we can wake up and maybe retry the failed work
-        if (!wm.isWorkInFlightMeetingTarget()) {
-            // though check if we have work awaiting retry
-            var lowestScheduledOpt = wm.getLowestRetryTime();
-            if (lowestScheduledOpt.isPresent()) {
-                // todo can sleep for less than this time? is this lower bound required? given that if we're starved - the failed work will most likely be selected? And even if not selected - then we will no longer be starved.
-                Duration retryDelay = options.getDefaultMessageRetryDelay();
-                // at min block for the retry time - retry time is not exact
-                Duration lowestScheduled = lowestScheduledOpt.get();
-                Duration timeBetweenCommits = getTimeBetweenCommits();
-                Duration effectiveRetryDelay = lowestScheduled.toMillis() < retryDelay.toMillis() ? retryDelay : lowestScheduled;
-                Duration result = timeBetweenCommits.toMillis() < effectiveRetryDelay.toMillis() ? timeBetweenCommits : effectiveRetryDelay;
-                log.debug("Not enough work in flight, while work is waiting to be retried - so will only sleep until next retry time of {} (lowestScheduled = {})", result, lowestScheduled);
-                return result;
-            }
-        }
-
-        //
-        Duration effectiveCommitAttemptDelay = getTimeToNextCommitCheck();
-        log.debug("Calculated next commit time in {}", effectiveCommitAttemptDelay);
-        return effectiveCommitAttemptDelay;
-    }
-
-    private boolean isIdlingOrRunning() {
-        return state == running || state == draining || state == paused;
-    }
-
-    protected boolean isTimeToCommitNow() {
-        updateLastCommitCheckTime();
-
-        Duration elapsedSinceLastCommit = this.lastCommitTime == null ? Duration.ofDays(1) : Duration.between(this.lastCommitTime, Instant.now());
-
-        boolean commitFrequencyOK = elapsedSinceLastCommit.compareTo(getTimeBetweenCommits()) > 0;
-        boolean lingerBeneficial = lingeringOnCommitWouldBeBeneficial();
-        boolean isCommandedToCommit = isCommandedToCommit();
-
-        boolean shouldDoANormalCommit = commitFrequencyOK && !lingerBeneficial;
-
-        boolean shouldCommitNow = shouldDoANormalCommit || isCommandedToCommit;
-
-        if (log.isDebugEnabled()) {
-            log.debug("Should commit this cycle? " +
-                    "shouldCommitNow? " + shouldCommitNow + " : " +
-                    "shouldDoANormalCommit? " + shouldDoANormalCommit + ", " +
-                    "commitFrequencyOK? " + commitFrequencyOK + ", " +
-                    "lingerBeneficial? " + lingerBeneficial + ", " +
-                    "isCommandedToCommit? " + isCommandedToCommit
-            );
-        }
-
-        return shouldCommitNow;
-    }
-
-    private int getNumberOfUserFunctionsQueued() {
-        return workerThreadPool.getQueue().size();
-    }
-
-    /**
-     * Under some conditions, waiting longer before committing can be faster
-     *
-     * @return true if waiting to commit would help performance
-     */
-    private boolean lingeringOnCommitWouldBeBeneficial() {
-        // work is waiting to be done
-        boolean workIsWaitingToBeCompletedSuccessfully = wm.workIsWaitingToBeProcessed();
-        // no work is currently being done
-        boolean workInFlight = wm.hasWorkInFlight();
-        // work mailbox is empty
-        boolean workWaitingInMailbox = !workMailbox.isEmpty();
-        boolean workWaitingToProcess = wm.hasIncompleteOffsets();
-        log.trace("workIsWaitingToBeCompletedSuccessfully {} || workInFlight {} || workWaitingInMailbox {} || !workWaitingToProcess {};",
-                workIsWaitingToBeCompletedSuccessfully, workInFlight, workWaitingInMailbox, !workWaitingToProcess);
-        boolean result = workIsWaitingToBeCompletedSuccessfully || workInFlight || workWaitingInMailbox || !workWaitingToProcess;
-
-        // todo disable - commit frequency takes care of lingering? is this outdated?
-        return false;
-    }
-
-    private Duration getTimeToNextCommitCheck() {
-        // draining is a normal running mode for the controller
-        if (isIdlingOrRunning()) {
-            Duration timeSinceLastCommit = getTimeSinceLastCheck();
-            Duration timeBetweenCommits = getTimeBetweenCommits();
-            @SuppressWarnings("UnnecessaryLocalVariable")
-            Duration minus = timeBetweenCommits.minus(timeSinceLastCommit);
-            return minus;
-        } else {
-            log.debug("System not {} (state: {}), so don't wait to commit, only a small thread yield time", running, state);
-            return Duration.ZERO;
-        }
-    }
-
-    private Duration getTimeSinceLastCheck() {
-        Instant now = clock.instant();
-        return Duration.between(lastCommitCheckTime, now);
-    }
-
-    /**
-     * Visible for testing
-     */
-    protected void commitOffsetsThatAreReady() throws TimeoutException, InterruptedException {
-        log.trace("Synchronizing on commitCommand...");
-        synchronized (commitCommand) {
-            log.debug("Committing offsets that are ready...");
-            committer.retrieveOffsetsAndCommit();
-            clearCommitCommand();
-            this.lastCommitTime = Instant.now();
-        }
-    }
-
-    private void updateLastCommitCheckTime() {
-        lastCommitCheckTime = Instant.now();
-    }
-
-    /**
-     * Early notify of work arrived.
-     * <p>
-     * Only wake up the thread if it's sleeping while polling the mail box.
-     *
-     * @see #processWorkCompleteMailBox
-     * @see #blockableControlThread
-     */
-    public void notifySomethingToDo() {
-        boolean noTransactionInProgress = !producerManager.map(ProducerManager::isTransactionCommittingInProgress).orElse(false);
-        if (noTransactionInProgress) {
-            log.trace("Interrupting control thread: Knock knock, wake up! You've got mail (tm)!");
-            interruptControlThread();
-        } else {
-            log.trace("Would have interrupted control thread, but TX in progress");
-        }
     }
 
     @Override
@@ -612,53 +157,6 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
 
     public void setLongPollTimeout(Duration ofMillis) {
         BrokerPollSystem.setLongPollTimeout(ofMillis);
-    }
-
-    /**
-     * Request a commit as soon as possible (ASAP), overriding other constraints.
-     */
-    public void requestCommitAsap() {
-        log.debug("Registering command to commit next chance");
-        synchronized (commitCommand) {
-            this.commitCommand.set(true);
-        }
-        notifySomethingToDo();
-    }
-
-    private boolean isCommandedToCommit() {
-        synchronized (commitCommand) {
-            return this.commitCommand.get();
-        }
-    }
-
-    private void clearCommitCommand() {
-        synchronized (commitCommand) {
-            if (commitCommand.get()) {
-                log.debug("Command to commit asap received, clearing");
-                this.commitCommand.set(false);
-            }
-        }
-    }
-
-    @Override
-    public void pauseIfRunning() {
-        if (this.state == State.running) {
-            log.info("Transitioning parallel consumer to state paused.");
-            this.state = State.paused;
-        } else {
-            log.debug("Skipping transition of parallel consumer to state paused. Current state is {}.", this.state);
-        }
-    }
-
-    @Override
-    public void resumeIfPaused() {
-        if (this.state == State.paused) {
-            log.info("Transitioning parallel consumer to state running.");
-            this.state = State.running;
-            notifySomethingToDo();
-        } else {
-            log.debug("Skipping transition of parallel consumer to state running. Current state is {}.", this.state);
-        }
     }
 
 }
